refactor: 동아리 모집 공고 도메인을 하위 패키지로 분리
