+++ src/main/java/in/koreatech/koin/domain/club/recruitment/dto/request/ClubRecruitmentModifyRequest.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.club.dto.request;
+package in.koreatech.koin.domain.club.recruitment.dto.request;
 
 import static com.fasterxml.jackson.databind.PropertyNamingStrategies.SnakeCaseStrategy;
 import static in.koreatech.koin.global.code.ApiResponseCode.*;
@@ -12,12 +12,10 @@ import com.fasterxml.jackson.annotation.JsonFormat;
 import com.fasterxml.jackson.databind.annotation.JsonNaming;
 
 import in.koreatech.koin.global.exception.CustomException;
-import in.koreatech.koin.domain.club.model.Club;
-import in.koreatech.koin.domain.club.model.ClubRecruitment;
 import io.swagger.v3.oas.annotations.media.Schema;
 
 @JsonNaming(value = SnakeCaseStrategy.class)
-public record ClubRecruitmentCreateRequest(
+public record ClubRecruitmentModifyRequest(
     @Schema(description = "모집 시작 기간", example = "2025-07-01", requiredMode = NOT_REQUIRED)
     @JsonFormat(pattern = "yyyy-MM-dd")
     LocalDate startDate,
@@ -29,13 +27,13 @@ public record ClubRecruitmentCreateRequest(
     @Schema(description = "상시 모집 여부", example = "false", requiredMode = REQUIRED)
     Boolean isAlwaysRecruiting,
 
-    @Schema(description = "모집 이미지", example = "https://bcsdlab.com/static/img/logo.d89d9cc.png", requiredMode = NOT_REQUIRED)
+    @Schema(description = "모집 이미지", example = "https://bcsdlab.com/static/img/logo.d89d9cc.png", requiredMode = REQUIRED)
     String imageUrl,
 
-    @Schema(description = "모집 상세 설명", example = "BCSD LAB 모집", requiredMode = NOT_REQUIRED)
+    @Schema(description = "모집 상세 설명", example = "BCSD LAB 모집", requiredMode = REQUIRED)
     String content
 ) {
-    public ClubRecruitmentCreateRequest {
+    public ClubRecruitmentModifyRequest {
         if (isAlwaysRecruiting && (startDate != null || endDate != null)) {
             throw CustomException.of(MUST_BE_NULL_RECRUITMENT_PERIOD);
         }
@@ -48,16 +46,5 @@ public record ClubRecruitmentCreateRequest(
                 throw CustomException.of(INVALID_RECRUITMENT_PERIOD);
             }
         }
-    }
-
-    public ClubRecruitment toEntity(Club club) {
-        return ClubRecruitment.builder()
-            .startDate(startDate)
-            .endDate(endDate)
-            .isAlwaysRecruiting(isAlwaysRecruiting)
-            .content(content)
-            .imageUrl(imageUrl)
-            .club(club)
-            .build();
     }
 }
