+++ src/main/java/in/koreatech/koin/domain/club/recruitment/model/ClubRecruitmentSubscription.java
@@ -1,9 +1,10 @@
-package in.koreatech.koin.domain.club.model;
+package in.koreatech.koin.domain.club.recruitment.model;
 
 import static jakarta.persistence.GenerationType.IDENTITY;
 import static lombok.AccessLevel.PROTECTED;
 
 import in.koreatech.koin.common.model.BaseEntity;
+import in.koreatech.koin.domain.club.model.Club;
 import in.koreatech.koin.domain.user.model.User;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
