+++ src/main/java/in/koreatech/koin/domain/club/recruitment/repository/ClubRecruitmentRepository.java
@@ -1,4 +1,4 @@
-package in.koreatech.koin.domain.club.repository;
+package in.koreatech.koin.domain.club.recruitment.repository;
 
 import java.util.Optional;
 
@@ -7,7 +7,7 @@ import org.springframework.data.repository.Repository;
 import in.koreatech.koin.global.code.ApiResponseCode;
 import in.koreatech.koin.global.exception.CustomException;
 import in.koreatech.koin.domain.club.model.Club;
-import in.koreatech.koin.domain.club.model.ClubRecruitment;
+import in.koreatech.koin.domain.club.recruitment.model.ClubRecruitment;
 
 public interface ClubRecruitmentRepository extends Repository<ClubRecruitment, Integer> {
