+++ src/test/java/in/koreatech/koin/unit/fixture/ClubBaseInfoFixture.java
@@ -2,7 +2,7 @@ package in.koreatech.koin.unit.fixture;
 
 import java.time.LocalDate;
 
-import in.koreatech.koin.domain.club.enums.ClubRecruitmentStatus;
+import in.koreatech.koin.domain.club.recruitment.enums.ClubRecruitmentStatus;
 import in.koreatech.koin.domain.club.model.ClubBaseInfo;
 
 public class ClubBaseInfoFixture {
